Update appCore submodule. And fixed crasher.
Update appCore submodule. And fixed crasher. Crasher was occuring when the user starts the fitness test task and leaves the app while on instruction step one or two. When the app restarts it is checking for a new step to inject into instruction step one or two in order to change the title. The fix was made by moving the adjustment into where the task steps are originally created.